@@ -1,18 +1,23 @@
 package com.example.application.views.list;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import com.example.application.data.entity.Contact;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.data.provider.ListDataProvider;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class ListViewTest {
+
+    static {
+        // Prevent Vaadin Development mode to launch browser window
+        System.setProperty("vaadin.launch-browser", "false");
+    }
 
     @Autowired
     private ListView listView;
@@ -24,11 +29,12 @@ public class ListViewTest {
 
         ContactForm form = listView.form;
 
-        Assert.assertFalse(form.isVisible());
+        assertFalse(form.isVisible());
         grid.asSingleSelect().setValue(firstContact);
-        Assert.assertTrue(form.isVisible());
-        Assert.assertEquals(firstContact.getFirstName(), form.firstName.getValue());
+        assertTrue(form.isVisible());
+        assertEquals(firstContact.getFirstName(), form.firstName.getValue());
     }
+
     private Contact getFirstItem(Grid<Contact> grid) {
         return( (ListDataProvider<Contact>) grid.getDataProvider()).getItems().iterator().next();
     }
