@@ -1,6 +1,5 @@
 package com.example.application;
 
-import com.vaadin.flow.component.dependency.NpmPackage;
 import com.vaadin.flow.component.page.AppShellConfigurator;
 import com.vaadin.flow.server.PWA;
 import com.vaadin.flow.theme.Theme;
@@ -9,11 +8,19 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
  * The entry point of the Spring Boot application.
+ *
+ * Use the @PWA annotation make the application installable on phones, tablets
+ * and some desktop browsers.
+ *
  */
 @SpringBootApplication
-@NpmPackage(value = "lumo-css-framework", version = "^4.0.10")
-@Theme("flowcrmtutorial")
-@PWA(name = "VaadinCRM", shortName = "CRM", offlinePath="offline.html", offlineResources = { "./images/offline.png"})
+@Theme(value = "flowcrmtutorial")
+@PWA(
+        name = "Vaadin CRM",
+        shortName = "CRM",
+        offlinePath="offline.html",
+        offlineResources = { "images/offline.png" }
+)
 public class Application implements AppShellConfigurator {
 
     public static void main(String[] args) {
