@@ -1,11 +1,11 @@
 package com.example.application.data.entity;
 
+import jakarta.annotation.Nullable;
+import jakarta.persistence.Entity;
+import jakarta.persistence.OneToMany;
+import jakarta.validation.constraints.NotBlank;
 import org.hibernate.annotations.Formula;
 
-import javax.annotation.Nullable;
-import javax.persistence.Entity;
-import javax.persistence.OneToMany;
-import javax.validation.constraints.NotBlank;
 import java.util.LinkedList;
 import java.util.List;
 
