@@ -1,17 +1,16 @@
 package com.example.application.data.repository;
 
 import com.example.application.data.entity.Contact;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ContactRepository extends JpaRepository<Contact, Integer> {
+public interface ContactRepository extends JpaRepository<Contact, Long> {
 
     @Query("select c from Contact c " +
-        "where lower(c.firstName) like lower(concat('%', :searchTerm, '%')) " +
-        "or lower(c.lastName) like lower(concat('%', :searchTerm, '%'))")
+            "where lower(c.firstName) like lower(concat('%', :searchTerm, '%')) " +
+            "or lower(c.lastName) like lower(concat('%', :searchTerm, '%'))")
     List<Contact> search(@Param("searchTerm") String searchTerm);
 }
