@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
+@Service 
 public class CrmService {
 
     private final ContactRepository contactRepository;
@@ -19,14 +19,14 @@ public class CrmService {
 
     public CrmService(ContactRepository contactRepository,
                       CompanyRepository companyRepository,
-                      StatusRepository statusRepository) {
+                      StatusRepository statusRepository) { 
         this.contactRepository = contactRepository;
         this.companyRepository = companyRepository;
         this.statusRepository = statusRepository;
     }
 
     public List<Contact> findAllContacts(String stringFilter) {
-        if (stringFilter == null || stringFilter.isEmpty()) {
+        if (stringFilter == null || stringFilter.isEmpty()) { 
             return contactRepository.findAll();
         } else {
             return contactRepository.search(stringFilter);
@@ -42,7 +42,7 @@ public class CrmService {
     }
 
     public void saveContact(Contact contact) {
-        if (contact == null) {
+        if (contact == null) { 
             System.err.println("Contact is null. Are you sure you have connected your form to the application?");
             return;
         }
