@@ -1,42 +1,41 @@
 package com.example.application.security;
 
-import java.util.Collections;
-
 import com.example.application.views.LoginView;
+import com.vaadin.flow.spring.security.VaadinWebSecurity;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 
-import com.vaadin.flow.spring.security.VaadinWebSecurity;
-
-@EnableWebSecurity
+@EnableWebSecurity // <1>
 @Configuration
-public class SecurityConfig extends VaadinWebSecurity {
+public class SecurityConfig extends VaadinWebSecurity { // <2>
 
-  private static class CrmInMemoryUserDetailsManager extends InMemoryUserDetailsManager {
-    public CrmInMemoryUserDetailsManager() {
-      createUser(new User("user",
-              "{noop}userpass",
-              Collections.singleton(new SimpleGrantedAuthority("ROLE_USER"))));
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.authorizeHttpRequests()
+                .requestMatchers("/images/*.png").permitAll();  // <3>
+        super.configure(http);
+        setLoginView(http, LoginView.class); // <4>
     }
-  }
 
-  @Override
-  protected void configure(HttpSecurity http) throws Exception {
-    // Authorize access to /images/ without authentication
-    http.authorizeRequests().antMatchers("/images/**").permitAll();
-    // Set default security policy that permits Vaadin internal requests and
-    // denies all other
-    super.configure(http);
-    setLoginView(http, LoginView.class, "/logout");
-  }
-
-  @Bean
-  public InMemoryUserDetailsManager userDetailsService() {
-    return new CrmInMemoryUserDetailsManager();
-  }
+    @Bean
+    public UserDetailsService users() {
+        UserDetails user = User.builder()
+                .username("user")
+                // password = password with this hash, don't tell anybody :-)
+                .password("{bcrypt}$2a$10$GRLdNijSQMUvl/au9ofL.eDwmoohzzS7.rmNSJZ.0FxO/BTk76klW")
+                .roles("USER")
+                .build();
+        UserDetails admin = User.builder()
+                .username("admin")
+                .password("{bcrypt}$2a$10$GRLdNijSQMUvl/au9ofL.eDwmoohzzS7.rmNSJZ.0FxO/BTk76klW")
+                .roles("USER", "ADMIN")
+                .build();
+        return new InMemoryUserDetailsManager(user, admin); // <5>
+    }
 }
