@@ -6,29 +6,33 @@ import com.vaadin.flow.component.charts.Chart;
 import com.vaadin.flow.component.charts.model.ChartType;
 import com.vaadin.flow.component.charts.model.DataSeries;
 import com.vaadin.flow.component.charts.model.DataSeriesItem;
+import com.vaadin.flow.component.html.Image;
 import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
+import com.vaadin.flow.theme.lumo.LumoUtility;
+import jakarta.annotation.security.PermitAll;
 
-import javax.annotation.security.PermitAll;
-
-@Route(value = "dashboard", layout = MainLayout.class)
-@PageTitle("Dashboard | Vaadin CRM")
 @PermitAll
+@Route(value = "dashboard", layout = MainLayout.class) // <1>
+@PageTitle("Dashboard | Vaadin CRM")
 public class DashboardView extends VerticalLayout {
     private final CrmService service;
 
-    public DashboardView(CrmService service) {
+    public DashboardView(CrmService service) { // <2>
         this.service = service;
         addClassName("dashboard-view");
-        setDefaultHorizontalComponentAlignment(Alignment.CENTER);
+        setDefaultHorizontalComponentAlignment(Alignment.CENTER); // <3>
+
         add(getContactStats(), getCompaniesChart());
     }
 
     private Component getContactStats() {
-        Span stats = new Span(service.countContacts() + " contacts");
-        stats.addClassNames("text-xl", "mt-m");
+        Span stats = new Span(service.countContacts() + " contacts"); // <4>
+        stats.addClassNames(
+            LumoUtility.FontSize.XLARGE,
+            LumoUtility.Margin.Top.MEDIUM);
         return stats;
     }
 
@@ -37,7 +41,7 @@ public class DashboardView extends VerticalLayout {
 
         DataSeries dataSeries = new DataSeries();
         service.findAllCompanies().forEach(company ->
-            dataSeries.add(new DataSeriesItem(company.getName(), company.getEmployeeCount())));
+            dataSeries.add(new DataSeriesItem(company.getName(), company.getEmployeeCount()))); // <5>
         chart.getConfiguration().setSeries(dataSeries);
         return chart;
     }
