@@ -9,8 +9,8 @@ import com.vaadin.flow.component.html.H1;
 import com.vaadin.flow.component.orderedlayout.FlexComponent;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
-import com.vaadin.flow.router.HighlightConditions;
 import com.vaadin.flow.router.RouterLink;
+import com.vaadin.flow.theme.lumo.LumoUtility;
 
 public class MainLayout extends AppLayout {
     private final SecurityService securityService;
@@ -23,28 +23,30 @@ public class MainLayout extends AppLayout {
 
     private void createHeader() {
         H1 logo = new H1("Vaadin CRM");
-        logo.addClassNames("text-l", "m-m");
+        logo.addClassNames(
+            LumoUtility.FontSize.LARGE,
+            LumoUtility.Margin.MEDIUM);
 
-        Button logout = new Button("Log out", e -> securityService.logout());
+        String u = securityService.getAuthenticatedUser().getUsername();
+        Button logout = new Button("Log out " + u, e -> securityService.logout()); // <2>
 
-        HorizontalLayout header = new HorizontalLayout(new DrawerToggle(), logo, logout);
+        var header = new HorizontalLayout(new DrawerToggle(), logo, logout);
 
         header.setDefaultVerticalComponentAlignment(FlexComponent.Alignment.CENTER);
-        header.expand(logo);
-        header.setWidth("100%");
-        header.addClassNames("py-0", "px-m");
+        header.expand(logo); // <4>
+        header.setWidthFull();
+        header.addClassNames(
+            LumoUtility.Padding.Vertical.NONE,
+            LumoUtility.Padding.Horizontal.MEDIUM);
 
-        addToNavbar(header);
+        addToNavbar(header); 
 
     }
 
     private void createDrawer() {
-        RouterLink listLink = new RouterLink("List", ListView.class);
-        listLink.setHighlightCondition(HighlightConditions.sameLocation());
-
         addToDrawer(new VerticalLayout(
-            listLink,
-            new RouterLink("Dashboard", DashboardView.class)
+                new RouterLink("List", ListView.class),
+                new RouterLink("Dashboard", DashboardView.class)
         ));
     }
 }
