@@ -3,6 +3,7 @@ package com.example.application.views.list;
 import com.example.application.data.entity.Company;
 import com.example.application.data.entity.Contact;
 import com.example.application.data.entity.Status;
+import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.Key;
@@ -15,24 +16,22 @@ import com.vaadin.flow.component.textfield.EmailField;
 import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.data.binder.BeanValidationBinder;
 import com.vaadin.flow.data.binder.Binder;
-import com.vaadin.flow.data.binder.ValidationException;
 import com.vaadin.flow.shared.Registration;
 
 import java.util.List;
 
-public class ContactForm extends FormLayout {
-  private Contact contact;
-
-  TextField firstName = new TextField("First name");
+public class ContactForm extends FormLayout { 
+  TextField firstName = new TextField("First name"); 
   TextField lastName = new TextField("Last name");
   EmailField email = new EmailField("Email");
   ComboBox<Status> status = new ComboBox<>("Status");
   ComboBox<Company> company = new ComboBox<>("Company");
-  Binder<Contact> binder = new BeanValidationBinder<>(Contact.class);
 
   Button save = new Button("Save");
   Button delete = new Button("Delete");
   Button close = new Button("Cancel");
+  // Other fields omitted
+  Binder<Contact> binder = new BeanValidationBinder<>(Contact.class);
 
   public ContactForm(List<Company> companies, List<Status> statuses) {
     addClassName("contact-form");
@@ -42,15 +41,16 @@ public class ContactForm extends FormLayout {
     company.setItemLabelGenerator(Company::getName);
     status.setItems(statuses);
     status.setItemLabelGenerator(Status::getName);
-    add(firstName,
+
+    add(firstName, 
         lastName,
         email,
         company,
         status,
-        createButtonsLayout()); 
+        createButtonsLayout());
   }
 
-  private HorizontalLayout createButtonsLayout() {
+  private Component createButtonsLayout() {
     save.addThemeVariants(ButtonVariant.LUMO_PRIMARY);
     delete.addThemeVariants(ButtonVariant.LUMO_ERROR);
     close.addThemeVariants(ButtonVariant.LUMO_TERTIARY);
@@ -58,28 +58,23 @@ public class ContactForm extends FormLayout {
     save.addClickShortcut(Key.ENTER);
     close.addClickShortcut(Key.ESCAPE);
 
-    save.addClickListener(event -> validateAndSave());
-    delete.addClickListener(event -> fireEvent(new DeleteEvent(this, contact)));
-    close.addClickListener(event -> fireEvent(new CloseEvent(this)));
+    save.addClickListener(event -> validateAndSave()); // <1>
+    delete.addClickListener(event -> fireEvent(new DeleteEvent(this, binder.getBean()))); // <2>
+    close.addClickListener(event -> fireEvent(new CloseEvent(this))); // <3>
 
-
-    binder.addStatusChangeListener(e -> save.setEnabled(binder.isValid()));
-
-    return new HorizontalLayout(save, delete, close); 
-  }
-
-  public void setContact(Contact contact) {
-    this.contact = contact;
-    binder.readBean(contact);
+    binder.addStatusChangeListener(e -> save.setEnabled(binder.isValid())); // <4>
+    return new HorizontalLayout(save, delete, close);
   }
 
   private void validateAndSave() {
-    try {
-      binder.writeBean(contact);
-      fireEvent(new SaveEvent(this, contact));
-    } catch (ValidationException e) {
-      e.printStackTrace();
+    if(binder.isValid()) {
+      fireEvent(new SaveEvent(this, binder.getBean())); // <6>
     }
+  }
+
+
+  public void setContact(Contact contact) {
+    binder.setBean(contact); // <1>
   }
 
   // Events
@@ -115,8 +110,17 @@ public class ContactForm extends FormLayout {
     }
   }
 
-  public <T extends ComponentEvent<?>> Registration addListener(Class<T> eventType,
-                                                                ComponentEventListener<T> listener) {
-    return getEventBus().addListener(eventType, listener);
+  public Registration addDeleteListener(ComponentEventListener<DeleteEvent> listener) {
+    return addListener(DeleteEvent.class, listener);
   }
+
+  public Registration addSaveListener(ComponentEventListener<SaveEvent> listener) {
+    return addListener(SaveEvent.class, listener);
+  }
+  public Registration addCloseListener(ComponentEventListener<CloseEvent> listener) {
+    return addListener(CloseEvent.class, listener);
+  }
+
+
 }
+
